授权书API

+++ shop-api/src/main/java/com/masiis/shop/api/service/user/UserCertificateService.java
@@ -54,8 +54,8 @@ public class UserCertificateService {
      * @Date 2016/3/17 0017 下午 4:30
      * 个人中心授权书
      */
-    public List<CertificateInfo> CertificateByUser(Integer userId) throws Exception{
-        List<CertificateInfo> certificateInfoList = certificateMapper.getCertificatesByUser(userId.longValue());
+    public List<CertificateInfo> CertificateByUser(Long userId) throws Exception{
+        List<CertificateInfo> certificateInfoList = certificateMapper.getCertificatesByUser(userId);
         if (certificateInfoList != null) {
             for (CertificateInfo certificateInfo : certificateInfoList) {
                 if (certificateInfo.getPid() != 0) {
